backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
openai==1.12.0